+++ lib/src/super_network_image_cache.dart
@@ -2,8 +2,11 @@ import 'package:super_network_image/src/super_network_image_cache_manager.dart';
 
 class SuperNetworkImageCache {
   /// Set the global cache duration for all images.
-  static void setGlobalCacheDuration(Duration duration) {
-    SuperNetworkImageCacheManager.instance.setGlobalCacheDuration(duration);
+  static void configure({String? cacheKey, Duration? duration}) {
+    SuperNetworkImageCacheManager.instance.configure(
+      cacheKey: cacheKey,
+      duration: duration,
+    );
   }
 
   /// Clear a specific image from the cache.