+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.2] - 2024-10-03
+
+- **Breaking Change** Removed the `setGlobalCacheDuration` method
+- **Added** a `configure` method to optionally set the global cache duration & the cache key
+
 ## [0.0.1] - 2024-10-03
 
 ### Added